Use WaitForDeletion helper in conditionalaccess package, move expand/flatten funcs into own source file
