+++ internal/services/conditionalaccess/conditional_access_policy_resource.go
@@ -17,6 +17,7 @@ import (
 	"github.com/manicminer/hamilton/odata"
 
 	"github.com/hashicorp/terraform-provider-azuread/internal/clients"
+	"github.com/hashicorp/terraform-provider-azuread/internal/helpers"
 	"github.com/hashicorp/terraform-provider-azuread/internal/tf"
 	"github.com/hashicorp/terraform-provider-azuread/internal/utils"
 	"github.com/hashicorp/terraform-provider-azuread/internal/validate"
@@ -515,348 +516,35 @@ func conditionalAccessPolicyResourceRead(ctx context.Context, d *schema.Resource
 
 func conditionalAccessPolicyResourceDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*clients.Client).ConditionalAccess.PoliciesClient
+	policyId := d.Id()
 
-	_, status, err := client.Get(ctx, d.Id(), odata.Query{})
+	_, status, err := client.Get(ctx, policyId, odata.Query{})
 	if err != nil {
 		if status == http.StatusNotFound {
-			log.Printf("[DEBUG] Conditional Access Policy with ID %q already deleted", d.Id())
+			log.Printf("[DEBUG] Conditional Access Policy with ID %q already deleted", policyId)
 			return nil
 		}
 
-		return tf.ErrorDiagPathF(err, "id", "Retrieving conditional access policy with ID %q", d.Id())
+		return tf.ErrorDiagPathF(err, "id", "Retrieving conditional access policy with ID %q", policyId)
 	}
 
-	status, err = client.Delete(ctx, d.Id())
+	status, err = client.Delete(ctx, policyId)
 	if err != nil {
-		return tf.ErrorDiagPathF(err, "id", "Deleting conditional access policy with ID %q, got status %d", d.Id(), status)
+		return tf.ErrorDiagPathF(err, "id", "Deleting conditional access policy with ID %q, got status %d", policyId, status)
 	}
 
-	log.Printf("[DEBUG] Waiting for conditional access policy %q to disappear", d.Id())
-	timeout, _ := ctx.Deadline()
-	stateConf := &resource.StateChangeConf{
-		Pending:                   []string{"Pending"},
-		Target:                    []string{"Deleted"},
-		Timeout:                   time.Until(timeout),
-		MinTimeout:                5 * time.Second,
-		ContinuousTargetOccurence: 5,
-		Refresh: func() (interface{}, string, error) {
-			client.BaseClient.DisableRetries = true
-			_, status, err := client.Get(ctx, d.Id(), odata.Query{})
+	if err := helpers.WaitForDeletion(ctx, func(ctx context.Context) (*bool, error) {
+		client.BaseClient.DisableRetries = true
+		if _, status, err := client.Get(ctx, policyId, odata.Query{}); err != nil {
 			if status == http.StatusNotFound {
-				return "stub", "Deleted", nil
+				return utils.Bool(false), nil
 			}
-			if err != nil {
-				return nil, "Error", err
-			}
-
-			return "stub", "Pending", nil
-		},
-	}
-	if _, err = stateConf.WaitForStateContext(ctx); err != nil {
-		return tf.ErrorDiagF(err, "waiting for deletion of conditional access policy with ID %q", d.Id())
+			return nil, err
+		}
+		return utils.Bool(true), nil
+	}); err != nil {
+		return tf.ErrorDiagF(err, "Waiting for deletion of conditional access policy with ID %q", policyId)
 	}
 
 	return nil
-}
-
-func flattenConditionalAccessConditionSet(in *msgraph.ConditionalAccessConditionSet) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"applications":        flattenConditionalAccessApplications(in.Applications),
-			"users":               flattenConditionalAccessUsers(in.Users),
-			"client_app_types":    tf.FlattenStringSlicePtr(in.ClientAppTypes),
-			"locations":           flattenConditionalAccessLocations(in.Locations),
-			"platforms":           flattenConditionalAccessPlatforms(in.Platforms),
-			"sign_in_risk_levels": tf.FlattenStringSlicePtr(in.SignInRiskLevels),
-			"user_risk_levels":    tf.FlattenStringSlicePtr(in.UserRiskLevels),
-		},
-	}
-}
-
-func flattenConditionalAccessApplications(in *msgraph.ConditionalAccessApplications) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"included_applications": tf.FlattenStringSlicePtr(in.IncludeApplications),
-			"excluded_applications": tf.FlattenStringSlicePtr(in.ExcludeApplications),
-			"included_user_actions": tf.FlattenStringSlicePtr(in.IncludeUserActions),
-		},
-	}
-}
-
-func flattenConditionalAccessUsers(in *msgraph.ConditionalAccessUsers) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"included_users":  tf.FlattenStringSlicePtr(in.IncludeUsers),
-			"excluded_users":  tf.FlattenStringSlicePtr(in.ExcludeUsers),
-			"included_groups": tf.FlattenStringSlicePtr(in.IncludeGroups),
-			"excluded_groups": tf.FlattenStringSlicePtr(in.ExcludeGroups),
-			"included_roles":  tf.FlattenStringSlicePtr(in.IncludeRoles),
-			"excluded_roles":  tf.FlattenStringSlicePtr(in.ExcludeRoles),
-		},
-	}
-}
-
-func flattenConditionalAccessLocations(in *msgraph.ConditionalAccessLocations) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"included_locations": tf.FlattenStringSlicePtr(in.IncludeLocations),
-			"excluded_locations": tf.FlattenStringSlicePtr(in.ExcludeLocations),
-		},
-	}
-}
-
-func flattenConditionalAccessPlatforms(in *msgraph.ConditionalAccessPlatforms) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"included_platforms": tf.FlattenStringSlicePtr(in.IncludePlatforms),
-			"excluded_platforms": tf.FlattenStringSlicePtr(in.ExcludePlatforms),
-		},
-	}
-}
-
-func flattenConditionalAccessGrantControls(in *msgraph.ConditionalAccessGrantControls) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"operator":                      in.Operator,
-			"built_in_controls":             tf.FlattenStringSlicePtr(in.BuiltInControls),
-			"custom_authentication_factors": tf.FlattenStringSlicePtr(in.CustomAuthenticationFactors),
-			"terms_of_use":                  tf.FlattenStringSlicePtr(in.TermsOfUse),
-		},
-	}
-}
-
-func flattenConditionalAccessSessionControls(in *msgraph.ConditionalAccessSessionControls) []interface{} {
-	if in == nil {
-		return []interface{}{}
-	}
-
-	applicationEnforceRestrictions := false
-	if in.ApplicationEnforcedRestrictions != nil {
-		applicationEnforceRestrictions = *in.ApplicationEnforcedRestrictions.IsEnabled
-	}
-
-	cloudAppSecurity := ""
-	if in.CloudAppSecurity != nil && in.CloudAppSecurity.CloudAppSecurityType != nil {
-		cloudAppSecurity = *in.CloudAppSecurity.CloudAppSecurityType
-	}
-
-	signInFrequency := 0
-	signInFrequencyPeriod := ""
-	if in.SignInFrequency != nil && in.SignInFrequency.Value != nil && in.SignInFrequency.Type != nil {
-		signInFrequency = int(*in.SignInFrequency.Value)
-		signInFrequencyPeriod = *in.SignInFrequency.Type
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"application_enforced_restrictions_enabled": applicationEnforceRestrictions,
-			"cloud_app_security_policy":                 cloudAppSecurity,
-			"sign_in_frequency":                         signInFrequency,
-			"sign_in_frequency_period":                  signInFrequencyPeriod,
-		},
-	}
-}
-
-func expandConditionalAccessConditionSet(in []interface{}) *msgraph.ConditionalAccessConditionSet {
-	if len(in) == 0 || in[0] == nil {
-		return nil
-	}
-
-	result := msgraph.ConditionalAccessConditionSet{}
-	config := in[0].(map[string]interface{})
-
-	applications := config["applications"].([]interface{})
-	users := config["users"].([]interface{})
-	clientAppTypes := config["client_app_types"].([]interface{})
-	locations := config["locations"].([]interface{})
-	platforms := config["platforms"].([]interface{})
-	signInRiskLevels := config["sign_in_risk_levels"].([]interface{})
-	userRiskLevels := config["user_risk_levels"].([]interface{})
-
-	result.Applications = expandConditionalAccessApplications(applications)
-	result.Users = expandConditionalAccessUsers(users)
-	result.ClientAppTypes = tf.ExpandStringSlicePtr(clientAppTypes)
-	result.Locations = expandConditionalAccessLocations(locations)
-	result.Platforms = expandConditionalAccessPlatforms(platforms)
-	result.SignInRiskLevels = tf.ExpandStringSlicePtr(signInRiskLevels)
-	result.UserRiskLevels = tf.ExpandStringSlicePtr(userRiskLevels)
-
-	return &result
-}
-
-func expandConditionalAccessApplications(in []interface{}) *msgraph.ConditionalAccessApplications {
-	if len(in) == 0 || in[0] == nil {
-		return nil
-	}
-
-	result := msgraph.ConditionalAccessApplications{}
-	config := in[0].(map[string]interface{})
-
-	includeApplications := config["included_applications"].([]interface{})
-	excludeApplications := config["excluded_applications"].([]interface{})
-	includeUserActions := config["included_user_actions"].([]interface{})
-
-	result.IncludeApplications = tf.ExpandStringSlicePtr(includeApplications)
-	result.ExcludeApplications = tf.ExpandStringSlicePtr(excludeApplications)
-	result.IncludeUserActions = tf.ExpandStringSlicePtr(includeUserActions)
-
-	return &result
-}
-
-func expandConditionalAccessUsers(in []interface{}) *msgraph.ConditionalAccessUsers {
-	if len(in) == 0 || in[0] == nil {
-		return nil
-	}
-
-	result := msgraph.ConditionalAccessUsers{}
-	config := in[0].(map[string]interface{})
-
-	includeUsers := config["included_users"].([]interface{})
-	excludeUsers := config["excluded_users"].([]interface{})
-	includeGroups := config["included_groups"].([]interface{})
-	excludeGroups := config["excluded_groups"].([]interface{})
-	includeRoles := config["included_roles"].([]interface{})
-	excludeRoles := config["excluded_roles"].([]interface{})
-
-	result.IncludeUsers = tf.ExpandStringSlicePtr(includeUsers)
-	result.ExcludeUsers = tf.ExpandStringSlicePtr(excludeUsers)
-	result.IncludeGroups = tf.ExpandStringSlicePtr(includeGroups)
-	result.ExcludeGroups = tf.ExpandStringSlicePtr(excludeGroups)
-	result.IncludeRoles = tf.ExpandStringSlicePtr(includeRoles)
-	result.ExcludeRoles = tf.ExpandStringSlicePtr(excludeRoles)
-
-	return &result
-}
-
-func expandConditionalAccessPlatforms(in []interface{}) *msgraph.ConditionalAccessPlatforms {
-	result := msgraph.ConditionalAccessPlatforms{}
-	if len(in) == 0 || in[0] == nil {
-		return &result
-	}
-
-	config := in[0].(map[string]interface{})
-
-	includePlatforms := config["included_platforms"].([]interface{})
-	excludePlatforms := config["excluded_platforms"].([]interface{})
-
-	result.IncludePlatforms = tf.ExpandStringSlicePtr(includePlatforms)
-	result.ExcludePlatforms = tf.ExpandStringSlicePtr(excludePlatforms)
-
-	return &result
-}
-
-func expandConditionalAccessLocations(in []interface{}) *msgraph.ConditionalAccessLocations {
-	if len(in) == 0 || in[0] == nil {
-		return nil
-	}
-
-	result := msgraph.ConditionalAccessLocations{}
-	config := in[0].(map[string]interface{})
-
-	includeLocations := config["included_locations"].([]interface{})
-	excludeLocations := config["excluded_locations"].([]interface{})
-
-	result.IncludeLocations = tf.ExpandStringSlicePtr(includeLocations)
-	result.ExcludeLocations = tf.ExpandStringSlicePtr(excludeLocations)
-
-	return &result
-}
-
-func expandConditionalAccessGrantControls(in []interface{}) *msgraph.ConditionalAccessGrantControls {
-	if len(in) == 0 || in[0] == nil {
-		return nil
-	}
-
-	result := msgraph.ConditionalAccessGrantControls{}
-	config := in[0].(map[string]interface{})
-
-	operator := config["operator"].(string)
-	builtInControls := config["built_in_controls"].([]interface{})
-	customAuthenticationFactors := config["custom_authentication_factors"].([]interface{})
-	termsOfUse := config["terms_of_use"].([]interface{})
-
-	result.Operator = &operator
-	result.BuiltInControls = tf.ExpandStringSlicePtr(builtInControls)
-	result.CustomAuthenticationFactors = tf.ExpandStringSlicePtr(customAuthenticationFactors)
-	result.TermsOfUse = tf.ExpandStringSlicePtr(termsOfUse)
-
-	return &result
-}
-
-func expandConditionalAccessSessionControls(in []interface{}, create bool) *msgraph.ConditionalAccessSessionControls {
-	result := msgraph.ConditionalAccessSessionControls{}
-
-	if create && (len(in) == 0 || in[0] == nil) {
-		return &result
-	}
-
-	// API doesn't accept boolean false values here in POST requests, it should be omitted instead
-	// When updating, omitting a setting doesn't change it, so we default to false
-	if !create {
-		result.ApplicationEnforcedRestrictions = &msgraph.ApplicationEnforcedRestrictionsSessionControl{
-			IsEnabled: utils.Bool(false),
-		}
-		result.CloudAppSecurity = &msgraph.CloudAppSecurityControl{
-			IsEnabled: utils.Bool(false),
-		}
-		result.SignInFrequency = &msgraph.SignInFrequencySessionControl{
-			IsEnabled: utils.Bool(false),
-		}
-	}
-
-	if len(in) == 0 || in[0] == nil {
-		return &result
-	}
-
-	config := in[0].(map[string]interface{})
-
-	result.ApplicationEnforcedRestrictions = &msgraph.ApplicationEnforcedRestrictionsSessionControl{
-		IsEnabled: utils.Bool(config["application_enforced_restrictions_enabled"].(bool)),
-	}
-
-	if cloudAppSecurity := config["cloud_app_security_policy"].(string); cloudAppSecurity != "" {
-		result.CloudAppSecurity = &msgraph.CloudAppSecurityControl{
-			IsEnabled:            utils.Bool(true),
-			CloudAppSecurityType: utils.String(cloudAppSecurity),
-		}
-	}
-
-	if signInFrequency := config["sign_in_frequency"].(int); signInFrequency > 0 {
-		result.SignInFrequency = &msgraph.SignInFrequencySessionControl{
-			IsEnabled: utils.Bool(true),
-			Type:      utils.String(config["sign_in_frequency_period"].(string)),
-			Value:     utils.Int32(int32(signInFrequency)),
-		}
-	}
-
-	// API doesn't accept all disabled settings on POST, instead should be an empty object
-	if create && !*result.ApplicationEnforcedRestrictions.IsEnabled && !*result.CloudAppSecurity.IsEnabled && !*result.SignInFrequency.IsEnabled {
-		return &msgraph.ConditionalAccessSessionControls{}
-	}
-
-	return &result
 }
